+++ demo/js/demo-nine.js
@@ -1,0 +1,74 @@
+const React = require('react');
+const ReactDOM = require('react-dom');
+const AriaModal = require('../../src/react-aria-modal');
+
+class DemoNine extends React.Component {
+  constructor(props) {
+    super(props);
+
+    this.state = {
+      modalActive: false
+    };
+
+    this.activateModal = this.activateModal.bind(this);
+    this.deactivateModal = this.deactivateModal.bind(this);
+    this.getApplicationNode = this.getApplicationNode.bind(this);
+  }
+
+  activateModal = () => {
+    this.setState({ modalActive: true });
+  };
+
+  deactivateModal = () => {
+    this.setState({ modalActive: false });
+  };
+
+  getApplicationNode = () => {
+    return document.getElementById('application');
+  };
+
+  render() {
+    const modal = this.state.modalActive
+      ? <AriaModal
+          titleText="demo nine"
+          onExit={this.deactivateModal}
+          initialFocus="#demo-nine-deactivate"
+          getApplicationNode={this.getApplicationNode}
+          underlayStyle={{ paddingTop: '2em' }}
+          scrollDisabled={false}
+        >
+          <div id="demo-nine-modal" className="modal">
+            <div className="modal-body">
+              <p>
+                Here is a modal
+                {' '}
+                <a href="#">with</a>
+                {' '}
+                <a href="#">some</a>
+                {' '}
+                <a href="#">focusable</a>
+                {' '}
+                parts.
+              </p>
+            </div>
+            <footer className="modal-footer">
+              <button id="demo-nine-deactivate" onClick={this.deactivateModal}>
+                deactivate modal
+              </button>
+            </footer>
+          </div>
+        </AriaModal>
+      : false;
+
+    return (
+      <div>
+        <button onClick={this.activateModal}>
+          activate modal
+        </button>
+        {modal}
+      </div>
+    );
+  }
+}
+
+ReactDOM.render(<DemoNine />, document.getElementById('demo-nine'));