+++ CHANGELOG.md
@@ -1,5 +1,17 @@
 # Changelog
 
+## 2.12.0
+
+- Add `focusTrapOptions` prop to pass options to the focus trap.
+
+## 2.11.1
+
+- Update `focus-trap-react` and `react-displace` to allow React 16 installation.
+
+## 2.11.0
+
+- Add `scrollDisabled` prop (default `true`), for edge cases when the document's scroll should not be disabled.
+
 ## 2.10.0
 
 - Add support for React 16.